feat: Add React and Lucide icons
This commit updates the project dependencies to include React and Lucide icons.
These libraries will be used for building the user interface of the InstaGenius application.

The `package.json` has been updated to reflect these new dependencies.
The `index.html` has been updated to include the necessary import maps for these new libraries.
The `ImageSelector.tsx` component has been updated to remove an unused import for `CheckCircle2` as it was not yet implemented.